refactor: rename routes to Router and use users-route.ts format
